moved actors out into their own files based on the entities they represent
